+++ src/App.jsx
@@ -1,36 +1,18 @@
-import React, { useState, useEffect } from "react";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import LoadingPage from "./components/LoadingPage";
-import MainPage from "./pages/MainPage";
-import Dashboard from "./pages/Dashboard";
+// App.jsx
+import React from 'react';
+import { Routes, Route } from 'react-router-dom';
+import LoginPage from './pages/LoginPage';
+import Dashboard from './pages/Dashboard'
 
 const App = () => {
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setIsLoading(false);
-    }, 3000);
-
-    return () => clearTimeout(timer); // Cleanup timer on unmount
-  }, []);
-
   return (
-    <Router>
-      <Routes>
-        <Route
-          path="/"
-          element={
-            isLoading ? (
-              <LoadingPage />
-            ) : (
-              <MainPage />
-            )
-          }
-        />
-        <Route path="/dashboard" element={<Dashboard />} />
-      </Routes>
-    </Router>
+    
+    <Routes>
+      <Route path="/" element={<LoginPage/>} />
+      <Route path='/dashboard' element={<Dashboard/>}/>
+     
+
+    </Routes>
   );
 };
