--- src/components/LoadingPage.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const LoadingPage = () => {
-  return (
-    <div className="flex items-center justify-center h-screen bg-gray-100">
-      <div className="w-16 h-16 border-4 border-blue-500 border-t-transparent rounded-full animate-spin"></div>
-    </div>
-  );
-};
-
-export default LoadingPage;
