+++ src/index.css
@@ -6,7 +6,7 @@
 
 body {
   font-family: 'Poppins', sans-serif;
-  
+  font-size: 15px;
 }
 
 input-container:hover .icon {
