+++ tailwind.config.js
@@ -7,9 +7,13 @@ export default {
   theme: {
     extend: {
 
+      fontFamily: {
+        poppins: ['Poppins', 'sans-serif'],
+      },
+
       color:{
         "dark-purple":"#081A51",
-        "light-white":"rgba(255,255,255,0.17)"
+        "light-white":"rgba(255,255,255,0.17)",
       },
       
       animation: {
